14: Certificate formatting
- Adding in logic to make sure that the formatting of the certificates
matches the format specified.

- Broke out program into modules to be consistent with snpguest.

- Fixed some linting issues

resolves #14
